Fix docstring, typing, refactor and move file
